+++ packages/yoastseo/spec/languageProcessing/helpers/sentence/getSentencesFromTreeSpec.js
@@ -8,7 +8,7 @@ describe( "test to get sentences from the tree", () => {
 	beforeEach( () => {
 		researcher = new EnglishResearcher();
 	} );
-	it( "test", () => {
+	it( "returns the sentences from paragraph and heading nodes", () => {
 		const paper = new Paper( "<p>A very intelligent cat loves their human. A dog is very cute.</p><h3>A subheading 3" +
 			"</h3>text text text<h4>A subheading 4</h4>more text." );
 		researcher.setPaper( paper );
@@ -88,6 +88,26 @@ describe( "test to get sentences from the tree", () => {
 					{ sourceCodeRange: { endOffset: 133, startOffset: 132 }, text: " " },
 					{ sourceCodeRange: { endOffset: 137, startOffset: 133 }, text: "text" },
 					{ sourceCodeRange: { endOffset: 138, startOffset: 137 }, text: "." },
+				],
+			},
+		] );
+	} );
+	it( "should not include sentences from non-paragraph and non-heading nodes", () => {
+		const paper = new Paper( "<p>A cute red panda</p><blockquote>The red panda (Ailurus fulgens), also known as the lesser panda," +
+			" is a small mammal native to the eastern Himalayas and southwestern China</blockquote>" );
+		researcher.setPaper( paper );
+		buildTree( paper, researcher );
+		expect( getSentencesFromTree( paper ) ).toEqual( [
+			{ sourceCodeRange: { endOffset: 19, startOffset: 3 },
+				text: "A cute red panda",
+				tokens: [
+					{ sourceCodeRange: { endOffset: 4, startOffset: 3 }, text: "A" },
+					{ sourceCodeRange: { endOffset: 5, startOffset: 4 }, text: " " },
+					{ sourceCodeRange: { endOffset: 9, startOffset: 5 }, text: "cute" },
+					{ sourceCodeRange: { endOffset: 10, startOffset: 9 }, text: " " },
+					{ sourceCodeRange: { endOffset: 13, startOffset: 10 }, text: "red" },
+					{ sourceCodeRange: { endOffset: 14, startOffset: 13 }, text: " " },
+					{ sourceCodeRange: { endOffset: 19, startOffset: 14 }, text: "panda" },
 				],
 			},
 		] );
